Bug 708006 - libsoup fails to close connection when message cancelled
We need to clear priv->reusable when we send a request, *even* if we're
only resending a request that was already tried once (as with NTLM
authentication). Otherwise, cancellation of the pending, restarted,
request will not properly close the connection as it should.

(cherry picked from commit b63d08471391c35b7c67d414f24fecf0d1810c74)