no-ssl-test: fix for GDummyTlsBackend changes
This was assuming that setting "ssl-use-system-ca-file" would fail
when using the dummy TLS backend, but the fact that it failed was just
a side effect of the fact that GDummyTlsBackend returned a different
GTlsDatabase every time you called
g_tls_backend_get_default_database(). Now that that's fixed, setting
"ssl-use-system-ca-file" does appear to work.

Of course, SSL won't work at all with the dummy backend, so it doesn't
really matter whether setting this property works or not, so just
remove that part of the test, so this will pass with either old or new
glib.