tests: Add resources/* to soup-test.gresource, add utility functions
Add the files in resources/ to soup-test.gresource, and add
soup_test_load_resource() to get the contents of one of them, and then
use this in places that were previously reading them by hand.

Except forms-test, which needs to pass a filename to curl, so have it
use index.txt instead of one of the resource files.

Now none of the tests access the resources/ directory at runtime.

Also add soup_test_get_index() to read index.txt rather than
reimplementing it in multiple test programs.