tests: ignore proxy settings
Some of the tests depend in various ways on the fact that they are not
going through a proxy, so make sure they really aren't (by forcing the
use of the dummy GProxyResolver). Likewise, pass "--noproxy *" to curl
in the tests that use it, to make sure it doesn't pick up $http_proxy,
etc.

Pointed out by Dirkjan Ochtman on IRC.