tests: update to handle G_PARAM_DEPRECATED warnings in new glib
SoupSession:ssl-ca-file is deprecated, but we were still using it some
tests for convenience, and in others because we need to test that it
still works right. Fix the former by using SoupSession:tls-database
instead, and fix the later by adjusting the property's GParamSpec's
flags around tests that need to test :ssl-ca-file.