cache: Clean up leaked resources when loading the cache
When soup_cache_dump() is not called for some reason, like a crash in
the application, the new cached resources are not referenced by the
index and then leaked in the cache directory. soup_cache_load() removes
all the resources found in the cache directory and not referenced by
the cached index.

https://bugzilla.gnome.org/show_bug.cgi?id=667682