cache: Do not add Age header to all responses
I think we were doing this because RFC 2616 says:

"When a response is generated from a cache entry, the cache MUST include a single Age
header field in the response with a value equal to the cache entry's current_age."

But it also says:

"The presence of an Age header field in a response implies that a
response is not first-hand. However, the converse is not true, since
the lack of an Age header field in a response does not imply that the
response is first-hand unless all caches along the request path are
compliant with HTTP/1.1"

We are always adding the Age header, even for responses sent after a
revalidation, in which case the response is first-hand. So at least
after a revalidation we should not include the Age header. However,
other browsers don't seem to include the Age header in any case, and the
first quote from RFC 2616 is not present in RFC 7234. So, the safest
approach that is compatible with other browsers and RFC compliant is to
never include the Age header manually in cache responses.

This avoids unnecessary conditional requests made when a resource is
considered expired because of the Age header added after a revalidation.

https://bugzilla.gnome.org/show_bug.cgi?id=741108