soup-message-io: document/rename broken GSource
The source returned by soup_message_io_get_source() depends on the
current state of the message, and so is not reusable. Document this,
and rename it to soup_message_io_get_oneshot_source().

Fix SoupClientInputStream, which was previously reusing its source.
Also, remove the "FIXME" code there, since the SoupSession bug it
referred to did eventually get fixed.