tests: fix bug in checking for a bug in xmlrpc-test
When porting to gtestutils, the check for "PHP is buggy" in
xmlrpc-test got messed up so that it triggered even when PHP wasn't
buggy, causing the test to get labelled as "skipped".

Since the PHP bug in question (https://bugs.php.net/45996) was fixed
years ago anyway, just remove the workaround for it.