cache: Do not revalidate resources that haven't expired even if must-revalidate is present
RFC2616 14.9.4

"When the must-revalidate directive is present in a response received by
a cache, that cache MUST NOT use the entry after it becomes stale to
respond to a subsequent request without first revalidating it with the origin server.
(I.e., the cache MUST do an end-to-end revalidation every time, if,
based solely on the origin server's Expires or max-age value, the cached
response is stale.)"

Update also the unit tests, since they were checking the wrong
behaviour.

https://bugzilla.gnome.org/show_bug.cgi?id=744468