Add SoupMessage flag SOUP_MESSAGE_IGNORE_CONNECTION_LIMITS
It ensures that a new connection will be created for the message if
there aren't any idle connections available and the current number of
connections have reached any of the limits. When a new dedicated
connection is created, it's automatically removed when the message is
unqueued, without switching to idle state to make sure it can't be
reused.

https://bugzilla.gnome.org/show_bug.cgi?id=744720