soup-message-io: fix up async body close semantics
add13ea0 fixed soup-message-io to do an async close of
io->body_ostream, so that in cases where it was necessary to write out
the final "0" chunk, and the stream was not immediately writable, it
wouldn't block.

However, it was emitting :wrote-body and potentially :finished before
the close finished in this case, which made it difficult to ensure
that the last request on a session was cleaned up correctly in the
test programs.

Fix this by adding in a new IO_STATE_BODY_FLUSH to do this step before
proceeding to IO_STATE_FINISHING. But also, to simplify things, just
do the sync close if we're not doing chunked encoding, since it's
guaranteed to not block in that case.

(Also revert the changes that had been made to continue-test for that
commit, since they are no longer needed.)