soup-socket: remove close-on-dispose
Setting close-on-dispose=FALSE on a socket would cause it to leak its
GSocket. Indeed, there is no way to avoid this, since freeing the
GSocket will always close it.

Work around the problem in a different way by having the steal()
methods steal the GSocket from the SoupSocket, and ensure that it gets
freed later along with the stolen GIOStream.