tests: Set default values for some variables
This eliminates a compiler warning about potentially using uninitialised
variables. It’s not actually possible for this happen (the variables are
only used on conditional paths with the same conditions as those where
they were previously set), but it’s easy enough to make this change to
shut the compiler up.

https://bugzilla.gnome.org/show_bug.cgi?id=748514