libsoup/soup-session.c: Don't Hardcode Paths on Windows
On Windows, use g_win32_get_package_installation_directory_of_module()
and g_build_filename() to acquire the localedir to find the translations
on Windows, instead of relying on the LOCALEDIR which is hardcoded at
compile time, so that the libsoup DLL can be relocated on Windows with its
translations.  Therefore, we define a simple DllMain() which is used to
acquire the proper HMODULE for the purpose.

https://bugzilla.gnome.org/show_bug.cgi?id=752952