soup-session.c: Fix Build on MinGW
MinGW requires that a prototype be declared for DllMain() before we
can use it, so define it here to fix the build, as the previous patch to
add this did not check for this part.  Also fix a build warning on the
DllMain() as a return BOOL was expected.

This patch was checked with Ignacio Quinteiro (Nacho) via e-mail
communication.