Fix runtime critical warning when loading an authenticated proxy resource from the disk cache
It happens when we load a resource that goes through an authenticated
proxy, and is cached in the disk. If the same resource is loaded again
the response is sent from the cache, and the auth manager tries to get
the proxy uri from the item connection. But there's no connection when
loading cached resources.

https://bugzilla.gnome.org/show_bug.cgi?id=756076