MSVC builds: Add config for MIT GSSAPI support
This adds support in the Visual Studio projects so that one can build
libSOUP with MIT GSSAPI support, given that MIT Kerberos is installed
beforehand, by using the configs with _MIT_GSSAPI in them.  The orginal
configs are still available and will still be the default Release and
Debug configs.

Also update config.h.win32.in so that it will be more consistent with the
entries checked for in the autotools-generated config.h.in, but the
LIBSOUP_HAVE_GSSAPI is activated only via the aforementioned project
configs.