Update translations for trivial error message change
The previous commit makes the "Could not listen on address %s, port
%d" string into a prefix rather than a complete error message, appending
": " to the translatable string.

In languages where the *other* error prefix string in libsoup ("Could
not import existing socket: ") is translated into a string still
ending in ": ", fix the "could not listen" translation as well by just
adding ": " to it. (an, as, bg, bn_IN, bs, ca, ca@valencia, da, de,
fr, fur, gu, hi, hu, ko, lt, lv, mr, nb, oc, or, pa, pl, pt, pt_BR, sk,
sr, sr@latin, sv, th, tr, vi)

In most languages where the "could not import" translation ends in ":"
but no space, but where the translation of "Invalid '%s' URI: %s"
*does* use a space after its colon, add the missing space to the
"could not import" translation, and then also update the "could not
listen" translation as above. (cs, es, eu, gl, id, it, ru, sl, te, uk)

In languages where the "could not import" translation ends in ":" but
no space, and it's not clear that the missing space is a bug, update
the "could not listen" translation to end in ":" but no space as well.
(he, kn, zh_CN)

In zh_HK and zh_TW, the "could not import" translation ends in a
full-width punctuation character, and so the "could not listen" string
has been updated to match.

This leaves untouched the languages that didn't have a translation of
the "could not import" string to compare to (be, en_GB, eo, et, ja,
ml, nl, ro, tg, ug, uz@cyrillic), or that translated it to something
not ending in a colon (el, ta).