Prefix signal ID enum entries with SIGNAL_ to avoid using EOF
glibc defines EOF as a macro, so we cannot use that as the name for a
signal ID enum entry, otherwise it will be preprocessed to some number
if we end up with the wrong header included, and compilation will fail.
Instead, prefix signal ID enum entries with ‘SIGNAL_’ in
SoupCacheClientInputStream and SoupClientInputStream.

https://bugzilla.gnome.org/show_bug.cgi?id=768731