SoupSession: give password to GSimpleProxyResolver
After setting the "proxy-uri" property on a SoupSession with embedded
basic auth (e.g. http://user:pass@example.com), libsoup does not make
use of the auth and ends up emitting an "authenticate" signal on
receiving 407. Calling soup_auth_anthenticate() with the same
credentials then works, but it should've made use of the auth from the
beginning to avoid the round trip.

This seems to be a regression from fb09bf93
(https://bugzilla.gnome.org/show_bug.cgi?id=680273), during the
transition to GSimpleProxyResolver. The issue is that libsoup uses
`soup_uri_to_string_internal`, which doesn't embed the password in the
resulting string.

The attached patch changes this, making sure to retain the previous
behaviour in all other cases.

https://bugzilla.gnome.org/show_bug.cgi?id=772932