Add API to disable using cached credentials for a particular SoupMessage
Add SOUP_MESSAGE_DO_NOT_USE_AUTH_CACHE new flag to tell the
SoupAuthManager that it shoudn't use cached credentials to authenticate a
particular message, and that credentials shouldn't be cached either
after a successful authentication.
When SOUP_MESSAGE_DO_NOT_USE_AUTH_CACHE flag is present for a message,
we never query the cached credentials on message starting callback, so
that the SoupAuth is not set there. The SoupAuth is now always set on
the message after authenticate, and the Authorization header updated
right before the message is re-queued when SOUP_MESSAGE_DO_NOT_USE_AUTH_CACHE
is used.
This patch also updates soup_message_set_auth() to no longer update the
Authorization header, but just set the SoupAuth handling the case of
setting the same SoupAuth twice.

https://bugzilla.gnome.org/show_bug.cgi?id=774033