Do not remove Authorization header unconditionally
This is a regression introduced in efcb377,
update_authorization_header() is always removing the Authorization, so
if the user added its own header it's ignored and authentication will
fail if we don't have cached credentials. This patch ensures the
behavior is the same as before efcb377 when flag
SOUP_MESSAGE_DO_NOT_USE_AUTH_CACHE is not present, by only removing the
Authorization header if the message has a SoupAuth when
update_authorization_header() is called.

https://bugzilla.gnome.org/show_bug.cgi?id=775882