auth: Fix async authentication when flag SOUP_MESSAGE_DO_NOT_USE_AUTH_CACHE is used
When the flag SOUP_MESSAGE_DO_NOT_USE_AUTH_CACHE is used, it's not possible
to successfully authenticate, and SOUP_STATUS_UNAUTHORIZED is always
returned even when soup_auth_autenticate was called with the right
credentials. This happens because we set the auth on the soup message right
after emitting the authenticate signal only if it was authenticated. If the
signal pauses the message, the auth will no longer be associated to the message,
and not cached either because flag SOUP_MESSAGE_DO_NOT_USE_AUTH_CACHE is
present. Since we always check if the auth returned by
soup_auth_get_message is ready before trying to use it, we can simply
always set the auth on the mssage right after emitting the authenticate
signal even if it was not authenticated yet. If it's eventually
authenticated then got-body callback will check it's ready to re-queue
the message as expected.

https://bugzilla.gnome.org/show_bug.cgi?id=777936