Visual Studio builds: Record version numbers in property sheets
Record the package version numbers in the property sheets in their own
fields so that they can be more easily reused.

We actually have them already in the parts of the command lines to
generate soup-version.h, but we want to move one step forward to refactor
how soup-version.h is generated on Visual Studio builds, as well as
generate the .pc files for the package.