soup-message-io: Do not fail when there's no empty line after headers
The spec says there should be an empty line (\r\n) between the response
headers and the body. However, some servers don't include the empty line
when the response doesn't have a body. This is causing several WebKit
tests to fail, because some of the imported w3c tests do not include
that empty line. Those tests pass in firefox, chromium and safari, so at
least those other browsers allow that.

https://bugzilla.gnome.org/show_bug.cgi?id=780352