SoupCache: fix setting default value for cache dir
The documentation mentions that a value of NULL can be used to let
SoupCache use its default, but code currently tries to call
g_file_test() on NULL in that case, which will cause a critical warning
to be generated.

Refactor how the cache dir default value is initialized to avoid that
issue.

https://bugzilla.gnome.org/show_bug.cgi?id=788452