soup-enum-types: Make sure items are exported properly
We need to decorate the symbols in the soup-enum-types.h.template with
GLIB_MKENUMS_EXTERN, and ensure that config.h is included in
soup-enum-types.c.template, so that the symbols in the generated
enumeration sources are exported from Visual Studio builds.

This will allow introspection builds to complete successfully on Visual
Studio.