meson: Build test utility library as static on MSVC builds
Visual Studio builds need __declspec(dllexport) directives to export
symbols and (more importantly) public variables.  Since this is a
library that is private to the test programs, build it as a static
library so that we don't have to go through decorating the function
prototypes and the extern variables in the headers.