soup-tld: check for trailing dots in domains
SoupCookieJar relies on SoupTLD to check whether cookies are for
toplevel domains. Up to this point, SoupTLD does not consider possible
trailing dots in domains, so a toplevel domain with a trailing dot
will be discarded (because it doesn't string-match any of the TLDs in
soup's copy of the public-suffix list). SoupCookieJar will then
erroneously believe that a TLD with a trailing dot is not a TLD and
will accept the cookie, which will later be passed to other domains
with the same TLD.

Fix this by discarding the trailing dot before looking up domains in
the public-suffix list. This way SoupTLD will work correctly with
domains with a trailing dot and SoupCookieJar will reject them if they
are TLDs.

https://gitlab.gnome.org/GNOME/libsoup/issues/5