meson: Use apachectl for obtaining the Apache configuration
For running some of the tests we need the Apache's httpd binary. As we want to
know more about its configuration we have to run it and parse the output. But
here is the first problem, because on Debian we can't run the binary unless the
/etc/apache2/envvars file is sourced, otherwise it ends with failure. The
recommended way to communicate with the Apache is the apachectl that passes
the arguments to httpd and also sources the envvars file. In the ideal world
we could use the apachectl to run the tests as well, but on Fedora any non
trivial call to it ends with the following error:
    Passing arguments to httpd using apachectl is no longer supported.

The summary is that for the configuration parsing we will use the apachectl,
but for running the tests we will use the httpd binary.

Closes: #7