soup-headers: add strict parameter parsing functions
These functions extend the existing parameter-parsing functions
but differ in that they return NULL if there are any duplicated
parameters. It is noted in their documentation that they
are not recommended to parse header fields that might contain
RFC5987-encoded parameters.

Add tests for the parameter parsing methods that cover the
different cases, including RFC5789-encoded parameters.