Clean up the usage of the SoupHSTSEnforcer::changed signal
The signal is emitted for any change in the enforcer policies.
Callers are responsible for verifying whether the changed policy
is permanent/session or not.

Check this in the DB enforcer, before modifying the database.

Document that the ::changed signal callbacks shouldn't modify
the policies received as parameters.

Remove comments that are no longer relevant.