connection: Fix check for remote disconnection when in idle state
A change in glib-networking has surfaced a tricky mistake in
soup_connection_get_state(). The goal of this code is to detect early if
the server has closed our connection, but the check didn't account for
the possibility that TLS-level data could be readable even if no
HTTP-level data is readable. We need to actually try reading from the
stream, rather than just checking if the socket is readable, since that
way any TLS-level data will be handled by GTlsConnection. We should
receive G_IO_ERROR_WOULD_BLOCK only if there is no HTTP-level data and
the connection is still open; otherwise, treat it as disconnected.

Many thanks to Dan Winship for explaining this confusing code after I
found it was problematic, and for proposing this clever solution.

Fixes #134