soup-auth-ntlm.c: Fix running on older Visual Studio CRT
This checks on Windows whether we have __USE_MINGW_ANSI_STDIO
defined for MinGW builds or whether we are building with Visual Studio
2015 or later, as sscanf(..., "%2hhx", ...) would either fail to build
or will fail to run correctly if either of these conditions do not hold
on Windows, as the stock Windows (pre-Visual Studio 2015) CRT does not
support the 'hh' modifier, as it is C99.

If the sscanf() implementation on Windows does not support the 'hh'
modifier, use a workaround so that we obtain the correct values we need
from sscanf().