ntlm: fix ntlmv2 authentication
This patch add the NTLMSSP_NEGOTIATE_EXTENDED_SESSIONSECURITY bit in the flags
of the NTLM type 3 response when using NTLMv2. This seems to be required for
some servers.

It also add 4 NULL bytes of padding at the end of the NTLM blob
(as described in http://davenport.sourceforge.net/ntlm.html)