WebSockets: message never finishes when handshake fails
When soup_websocket_client_verify_handshake() returns TRUE, the
message connection is stolen and soup_message_io_steal() is called for
the message, making the message to move to FINISHING state. However,
when it returns FALSE, the message stays in RUNNING state forever. We
should call soup_message_io_finished() in that case to ensure the
messages transitions to FINISHING state.