SoupDate: Move from GTimeVal to glibc APIs in soup_date_to_time_t()
GTimeVal is deprecated, so let's stop using it. Also add a test
for soup_date_to_time_t() which was missing.

The fixes in this change use time_t-specific APIs from libc only.
As such, they are not necessarily Y2038-safe depending on the
architecture. time_t is not safe anyway so there's no point in
trying to be. This method will be deprecated in the next release
cycle anyway.