WebSockets: document how extensions should treat buffers in case of error
In case of error while processing an incoming or outgoing message, an
extension should return NULL as is customary. This way, in case of
error, the caller doesn't need to worry about having to deal with
droppign the reference to the return value if non-NULL.

Make this clear in the class documentation. Also remove an unnecessary
g_bytes_unref() call from the send_message() error handling.