soup-session: fix race in soup_session_real_kick_queue()
If the current context in soup_session_real_kick_queue() is different
from item->async_context, the idle_run_queue() can be called before
the item->async_pending is set to TRUE.

If that happens, the item queue gets stucked because the item is not
able to finish -- item->async_pending is TRUE but there's no pending
source that would actually run the queue.

To fix this, add a hash table of currently pending contexts
and protect it with a mutex. With this change, item->async_pending
becomes obsolete.

Signed-off-by: Jakub Janků <jjanku@redhat.com>