test-utils: Clarify meaning of an environment variable
SOUP_TESTS_IN_MAKE_CHECK used to be used for the Autotools build system.
I mistakenly thought it was a way to skip the Apache-dependent tests
during `make check`, but in fact the Autotools build system used to
start a single instance of Apache, then run all the tests against that
single instance, and finally shut it down.

This mechanism is currently unused, but resurrecting it might be one way
to avoid GNOME/libsoup#175.