soup-message-io: Add sysprof profiling support for HTTP connections
This adds very basic support for dumping HTTP connection information to
sysprof, if the process is being run under a sysprof session.

See https://gitlab.gnome.org/GNOME/sysprof/-/issues/43 for plans of how
this could be expanded in future. This is just a starting point.

The code in this commit dumps a message to the sysprof capture which
includes the URI, total time for the connection (request + response),
and the amount of data transferred in the request and response.

It adds an optional dependency on `libsysprof-capture-4.a`, and a
subproject for building that if it’s not available on the system.

Signed-off-by: Philip Withnall <withnall@endlessm.com>