Move authenticate signal from SoupSession to SoupMessage
In case of tunneling the authenticate signal is emitted in the original
message, since the tunneling message is an internal implementation
detail. The authenticate signal is now true handled, and SoupAuthManager
automatically pauses and unpauses the message when handled
asynchronously. soup_auth_cancel() has been added to complete an async
authentication when no credentials are provided.