tests: add a autobahn test client
Autobahn is a WebSockets test suite that allows us to test libsoup's
implementation. With the files added it should be possible to start
running WebSocket tests in the CI in gitlab, but it will probably
still need some further work to integrate it nicely.

autobahn-server.sh will start a docker container that runs the
Autobahn fuzzing server. It uses fuzzingserver.json for the server
settings.

soup-autobahn-test-client is a simple client that can run Autobahn
tests by connecting to said server.