logger: ensure request cookies are shown in the log
Cookies feature includes the cookies request information in a
SoupMessage::starting callback, same callback used by the logger to
print the request information. If the cookies feature is created before
the logger, the starting logger callback is called before the cookies
one, and the cookies header hasn't been added yet. Use
g_signal_connect_after() in the logger to ensure the callback is called
after all other features have updated the message in the starting
signal.